Extracted heroimage to singular component, played around with header details
