+++ css/style.css
@@ -74,10 +74,25 @@ li {
 	color: white;
 	font-weight: 700;
 	font-size: 4rem;
+}
+
+
+.hero-image-text-container {
 	position: absolute;
-	top: 50%;
-	left: 50%;
-	transform: translate(-50%, -50%);
+    left: 50%;
+    top: 50%;
+    transform: translate(-50%, -50%);
+}
+
+.header-stripes {
+	position: absolute;
+	top: 16;
+	left: -37;
+    z-index: -4;
+	width: 150px;
+	height: 46;
+  	background-image: linear-gradient(45deg, #f3bf38 25%, rgba(255,0,0,0) 25%, rgba(255,0,0,0) 50%, #f3bf38 50%, #f3bf38 75%, rgba(255,0,0,0) 75%, rgba(255,0,0,0) 100%);
+	background-size: 28.28px 28.28px;
 }
 
 .header-image-container {
